+++ internal/knowledge/extract_fs_bench_test.go
@@ -76,6 +76,6 @@ func BenchmarkAssembleDocs(b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		_ = assembleDocs(root, spells)
+		_ = assembleDocs(root, spells, nil)
 	}
 }